[Myles] Makefile will copy config file to expected location
The Makefile will place a copy of the config file into the place it is expected. this will help with speeding up testing time automating the copying process

I will one day replace this with an auto generating solution so that even if an end user deleted it accidentally it is regenerated!

Myles